fix: separate context, provider, and hook into individual files
