+++ src/context/GameProvider.tsx
@@ -1,0 +1,12 @@
+import { useState, type ReactNode } from 'react';
+import { GameContext } from './GameContext';
+
+export function GameProvider({ children }: { children: ReactNode }) {
+  const [pseudo, setPseudo] = useState('');
+
+  return (
+    <GameContext.Provider value={{ pseudo, setPseudo }}>
+      {children}
+    </GameContext.Provider>
+  );
+}
