+++ src/pages/Home.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 import { Button } from '../components/Button';
-import { useGame } from '../context/GameContext';
+import { useGame } from '../hooks/useGame';
 
 export function Home() {
   const navigate = useNavigate();
